Added no_std feature switch.
Using this feature switch will also turn on alloc feature switch which
is used to determine if the collections used in the library should be
imported from the alloc library or not.

Most notable change in no_std is that instead of a HashMap and Set
a BTreeMap and Set are used.

Other squashed changes:
- Added cfg_attr for switching on nightly alloc feature
- Added nightly travis test with no_std feature enabled
- Fixed travis.yml syntax
- Use fixedbitset with no_std feature in no_std mode
- Run cargo fmt